Make classes easier to understand
